+++ src/kernel_long_gravity.h
@@ -46,22 +46,22 @@
 struct chi_derivatives {
 
   /*! 0th order derivative \f$\chi(r,r_s)\f$ */
-  float chi_0;
+  double chi_0;
 
   /*! 1st order derivative \f$\partial_{r}\chi(r,r_s)\f$ */
-  float chi_1;
+  double chi_1;
 
   /*! 2nd order derivative \f$\partial_{rr}\chi(r,r_s)\f$ */
-  float chi_2;
+  double chi_2;
 
   /*! 3rd order derivative \f$\partial_{rrr}\chi(r,r_s)\f$ */
-  float chi_3;
+  double chi_3;
 
   /*! 4th order derivative \f$\partial_{rrrr}\chi(r,r_s)\f$ */
-  float chi_4;
+  double chi_4;
 
   /*! 5th order derivative \f$\partial_{rrrrr}\chi(r,r_s)\f$ */
-  float chi_5;
+  double chi_5;
 };
 
 /**
@@ -73,32 +73,32 @@ struct chi_derivatives {
  * @param derivs (return) The computed #chi_derivatives.
  */
 __attribute__((always_inline)) INLINE static void kernel_long_grav_derivatives(
-    const float r, const float r_s_inv, struct chi_derivatives *const derivs) {
+    const double r, const double r_s_inv, struct chi_derivatives *const derivs) {
 
 #ifdef GADGET2_LONG_RANGE_CORRECTION
 
   /* Powers of u=r/2r_s */
-  const float u = 0.5f * r * r_s_inv;
-  const float u2 = u * u;
-  const float u3 = u2 * u;
-  const float u4 = u3 * u;
+  const double u = 0.5 * r * r_s_inv;
+  const double u2 = u * u;
+  const double u3 = u2 * u;
+  const double u4 = u3 * u;
 
   /* Powers of (1/r_s) */
-  const float r_s_inv2 = r_s_inv * r_s_inv;
-  const float r_s_inv3 = r_s_inv2 * r_s_inv;
-  const float r_s_inv4 = r_s_inv3 * r_s_inv;
-  const float r_s_inv5 = r_s_inv4 * r_s_inv;
+  const double r_s_inv2 = r_s_inv * r_s_inv;
+  const double r_s_inv3 = r_s_inv2 * r_s_inv;
+  const double r_s_inv4 = r_s_inv3 * r_s_inv;
+  const double r_s_inv5 = r_s_inv4 * r_s_inv;
 
   /* Derivatives of \chi */
   derivs->chi_0 = approx_erfcf(u);
   derivs->chi_1 = -r_s_inv;
   derivs->chi_2 = r_s_inv2 * u;
-  derivs->chi_3 = -r_s_inv3 * (u2 - 0.5f);
-  derivs->chi_4 = r_s_inv4 * (u3 - 1.5f * u);
-  derivs->chi_5 = -r_s_inv5 * (u4 - 3.f * u2 + 0.75f);
+  derivs->chi_3 = -r_s_inv3 * (u2 - 0.5);
+  derivs->chi_4 = r_s_inv4 * (u3 - 1.5 * u);
+  derivs->chi_5 = -r_s_inv5 * (u4 - 3. * u2 + 0.75);
 
-  const float one_over_sqrt_pi = ((float)(M_2_SQRTPI * 0.5));
-  const float common_factor = one_over_sqrt_pi * expf(-u2);
+  const double one_over_sqrt_pi = ((double)(M_2_SQRTPI * 0.5));
+  const double common_factor = one_over_sqrt_pi * exp(-u2);
 
   /* Multiply in the common factors */
   derivs->chi_1 *= common_factor;
@@ -110,38 +110,37 @@ __attribute__((always_inline)) INLINE static void kernel_long_grav_derivatives(
 #else
 
   /* Powers of 2/r_s */
-  const float c0 = 1.f;
-  const float c1 = 2.f * r_s_inv;
-  const float c2 = c1 * c1;
-  const float c3 = c2 * c1;
-  const float c4 = c3 * c1;
-  const float c5 = c4 * c1;
+  const double c0 = 1;
+  const double c1 = 2. * r_s_inv;
+  const double c2 = c1 * c1;
+  const double c3 = c2 * c1;
+  const double c4 = c3 * c1;
+  const double c5 = c4 * c1;
 
   /* 2r / r_s */
-  const float x = c1 * r;
+  const double x = c1 * r;
 
   /* e^(2r / r_s) */
-  const float exp_x = expf(x);  // good_approx_expf(x);
+  const double exp_x = exp(x);  // good_approx_expf(x);
 
   /* 1 / alpha(w) */
-  const float a_inv = 1.f + exp_x;
+  const double a_inv = 1. + exp_x;
 
   /* Powers of alpha */
-  const float a1 = 1.f / a_inv;
-  const float a2 = a1 * a1;
-  const float a3 = a2 * a1;
-  const float a4 = a3 * a1;
-  const float a5 = a4 * a1;
-  const float a6 = a5 * a1;
+  const double a1 = 1. / a_inv;
+  const double a2 = a1 * a1;
+  const double a3 = a2 * a1;
+  const double a4 = a3 * a1;
+  const double a5 = a4 * a1;
+  const double a6 = a5 * a1;
 
   /* Derivatives of \chi */
-  derivs->chi_0 = -2.f * exp_x * c0 * a1 + 2.f;
-  derivs->chi_1 = -2.f * exp_x * c1 * a2;
-  derivs->chi_2 = -2.f * exp_x * c2 * (2.f * a3 - a2);
-  derivs->chi_3 = -2.f * exp_x * c3 * (6.f * a4 - 6.f * a3 + a2);
-  derivs->chi_4 = -2.f * exp_x * c4 * (24.f * a5 - 36.f * a4 + 14.f * a3 - a2);
-  derivs->chi_5 = -2.f * exp_x * c5 *
-                  (120.f * a6 - 240.f * a5 + 150.f * a4 - 30.f * a3 + a2);
+  derivs->chi_0 = -2. * exp_x * c0 * a1 + 2.;
+  derivs->chi_1 = -2. * exp_x * c1 * a2;
+  derivs->chi_2 = -2. * exp_x * c2 * (2. * a3 - a2);
+  derivs->chi_3 = -2. * exp_x * c3 * (6. * a4 - 6. * a3 + a2);
+  derivs->chi_4 = -2. * exp_x * c4 * (24. * a5 - 36. * a4 + 14. * a3 - a2);
+  derivs->chi_5 = -2. * exp_x * c5 * (120. * a6 - 240. * a5 + 150. * a4 - 30. * a3 + a2);
 #endif
 }
